# yamls/openapi.yaml
swagger: '2.0'
host: api.example.com
schemes:
  - https
basePath: /xukele/v1
info:
  description: API description in Markdown.
  title: Sample API
  version: 1.0.0
tags:
  - name: KeleDemo
    description: demo接口
paths:
  /kele:
    get:
      tags:
        - KeleDemo
      description: 查询用户信息
      operationId: queryUserInfoById
      parameters:
        - description: 用户id
          in: query
          name: id
          required: true
          format: int64
          type: integer
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserInfoRsp'
      x-accepts: application/json
      produces:
        - application/json
    post:
      tags:
        - KeleDemo
      x-accepts: application/json
      description: 保存用户信息
      operationId: saveUserInfo
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SaveUserInfoRsp'
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/UserInfoData'
      consumes:
        - application/json
      produces:
        - application/json
  /time:
    get:
      description: 测试桥接
      tags:
        - KeleDemo
      operationId: testBridge
      parameters:
        - in: query
          name: uid
          required: true
          format: int64
          type: integer
        - in: query
          name: appId
          required: true
          type: string
      responses:
        '200':
          description: 时间
          schema:
            properties:
              code:
                type: integer
              time:
                type: integer
      produces:
        - application/json

definitions:
  SaveUserInfoRsp:
    $ref: '#/definitions/KeleResult'
  KeleResult:
    properties:
      code:
        description: 业务码
        type: integer
      message:
        description: 业务信息
        type: string
    type: object
  UserInfoData:
    properties:
      username:
        type: string
      age:
        type: integer
      skills:
        items:
          type: string
        type: array
    type: object
  UserInfoRsp:
    properties:
      ret:
        $ref: '#/definitions/KeleResult'
      data:
        $ref: '#/definitions/UserInfoData'
    type: object